refactor rasc
