+++ CODEOWNERS
@@ -1018,6 +1018,7 @@ build.json @home-assistant/supervisor
 /tests/components/random/ @fabaff
 /homeassistant/components/rapt_ble/ @sairon
 /tests/components/rapt_ble/ @sairon
+/homeassistant/components/rasc/ @kwang1012
 /homeassistant/components/raspberry_pi/ @home-assistant/core
 /tests/components/raspberry_pi/ @home-assistant/core
 /homeassistant/components/rdw/ @frenck @joostlek
