+++ homeassistant/components/automation/__init__.py
@@ -62,7 +62,11 @@ from homeassistant.helpers.integration_platform import (
     async_process_integration_platform_for_component,
 )
 from homeassistant.helpers.issue_registry import IssueSeverity, async_create_issue
-from homeassistant.helpers.rascalscheduler import RoutineEntity
+from homeassistant.helpers.rascalscheduler import (
+    RoutineEntity,
+    dag_operator,
+    get_rascal_scheduler,
+)
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.script import (
     ATTR_CUR,
@@ -92,11 +96,6 @@ from homeassistant.helpers.trigger import (
 )
 from homeassistant.helpers.typing import ConfigType
 from homeassistant.loader import bind_hass
-from homeassistant.rascalscheduler import (
-    dag_operator,
-    get_rascal_scheduler,
-    setup_rascal_scheduler_entity,
-)
 from homeassistant.util.dt import parse_datetime
 
 from .config import AutomationConfig
@@ -259,7 +258,7 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
     )
 
     # Set up rascal scheduler component
-    setup_rascal_scheduler_entity(hass)
+    # setup_rascal_scheduler_entity(hass)
 
     # Wait until script component is loaded
     # Convert automation entity into DAG
