+++ homeassistant/components/light/__init__.py
@@ -1077,7 +1077,7 @@ class LightEntity(ToggleEntity):
         """Flag supported features."""
         return self._attr_supported_features
 
-    async def async_get_action_target_state(
+    def async_get_action_target_state(
         self, action: dict[str, Any]
     ) -> dict[str, Any] | None:
         """Return expected state when action is complete.
@@ -1124,9 +1124,7 @@ class LightEntity(ToggleEntity):
 
             return match
 
-        target: dict[str, Any] = (
-            await super().async_get_action_completed_state(action) or {}
-        )
+        target: dict[str, Any] = super().async_get_action_target_state(action) or {}
 
         service_data = action[CONF_SERVICE_DATA]
         supported_color_modes = self._light_internal_supported_color_modes
