+++ homeassistant/components/rasc/const.py
@@ -1,0 +1,12 @@
+"""Constants for the rasc integration."""
+import logging
+
+DOMAIN = "rasc"
+
+RASC_ACK = "ack"
+RASC_START = "start"
+RASC_COMPLETE = "complete"
+RASC_RESPONSE = "rasc_response"
+RASC_SCHEDULED = "scheduled"
+
+LOGGER = logging.getLogger(__package__)
