+++ homeassistant/components/rasc/manifest.json
@@ -1,0 +1,7 @@
+{
+  "domain": "rasc",
+  "name": "RASC Abstraction",
+  "codeowners": ["@kwang1012"],
+  "documentation": "https://github.com/kwang1012/home-assistant-core",
+  "iot_class": "local_push"
+}
