+++ homeassistant/helpers/restore_state.py
@@ -350,7 +350,3 @@ class RestoreEntity(Entity):
         Implemented by platform classes.
         """
         return None
-
-    # async def async_get_action_completed_state(self, action: str | None) -> str | None:
-    #     """Return expected state when action is complete."""
-    #     return None
